% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dAktorzy.R
\name{dAktorzy}
\alias{dAktorzy}
\title{Liczenie odleglosci dla aktorow}
\usage{
dAktorzy(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\value{
Funkcja zwraca wartosc liczbowa opisujaca odleglosc miedzy filmami.
}
\description{
Funkcja \code{dAktorzy} liczy odleglosc miedzy aktorami dwoch filmow z bazy
}
\details{
Funkcja wyciaga ID dwoch filmow z bazy. Jesli ktorys z nich nie ma aktorow odleglosc
miedzy nimi wynosi 0.8, jesli tak nie jest miara jest liczona zgodnie z funkcja PorownajWektory.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
dAktorzy(4000,2461,"BazaFilmow.sql")
dAktorzy(23456,2461,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

