% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dGatunki.R
\name{dGatunki}
\alias{dGatunki}
\title{Liczenie odleglosci dla gatunkow}
\usage{
dGatunki(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\value{
Funkcja zwraca wartosc liczbowa opisujaca odleglosc miedzy filmami.
}
\description{
Funkcja \code{dGatunki} liczy odleglosc miedzy gatunkami dwoch filmow z bazy
}
\details{
Funkcja wyciaga ID dwoch filmow z bazy. Jesli ktorys z nich nie ma gatunkow odleglosc miedzy
nimi wynosi 0.8, jesli tak nie jest miara jest liczona zgodnie z funkcja PorownajWektory.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
dGatunki(4000,2461,"BazaFilmow.sql")
dGatunki(23456,2461,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

