% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dOcenaIMDB.R
\name{dOcenaIMDB}
\alias{dOcenaIMDB}
\title{Liczenie odleglosci dla ogolnych ocen}
\usage{
dOcenaIMDB(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\value{
Funkcja zwraca wartosc liczbowa opisujaca odleglosc miedzy filmami.
}
\description{
Funkcja \code{dOcenaIMDB} liczy odleglosc miedzyogolnymi ocenami dwoch filmow z bazy
}
\details{
Funkcja wyciaga ID dwoch filmow z bazy. Jesli ktorys z nich nie maogolnych ocen odleglosc
miedzy nimi wynosi 0.5, jesli tak nie jest miara jest liczona zgodnie z funkcja PorownajOceny.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
dOcenaIMDB(4000,2461,"BazaFilmow.sql")
dOcenaIMDB(23456,2461,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

