% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dOcenaKobiety30_44.R
\name{dOcenaKobiety30_44}
\alias{dOcenaKobiety30_44}
\title{Liczenie odleglosci dla  ocen kobiet od 30 do 44 lat}
\usage{
dOcenaKobiety30_44(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\value{
Funkcja zwraca wartosc liczbowa opisujaca odleglosc miedzy filmami.
}
\description{
Funkcja \code{dOcenaKobiety30_44} liczy odleglosc miedzy ocenami kobiet od 30 do 44 lat dla
dwoch filmow z bazy
}
\details{
Funkcja wyciaga ID dwoch filmow z bazy. Jesli ktorys z nich nie ma ocen kobiet od 30 do 44 lat
odleglosc miedzy nimi wynosi 0.5, jesli tak nie jest miara jest liczona zgodnie z funkcja PorownajOceny.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
dOcenaKobiety30_44(4000,2461,"BazaFilmow.sql")
dOcenaKobiety30_44(23456,2461,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

