% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dOcenaKobietyPonizej18.R
\name{dOcenaKobietyPonizej18}
\alias{dOcenaKobietyPonizej18}
\title{Liczenie odleglosci dla  ocen kobiet ponizej 18 lat}
\usage{
dOcenaKobietyPonizej18(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\value{
Funkcja zwraca wartosc liczbowa opisujaca odleglosc miedzy filmami.
}
\description{
Funkcja \code{dOcenaKobietyPonizej18} liczy odleglosc miedzy ocenami kobiet ponizej 18 lat dla
dwoch filmow z bazy
}
\details{
Funkcja wyciaga ID dwoch filmow z bazy. Jesli ktorys z nich nie ma ocen kobiet ponizej 18 lat
odleglosc miedzy nimi wynosi 0.5, jesli tak nie jest miara jest liczona zgodnie z funkcja PorownajOceny.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
dOcenaKobietyPonizej18(4000,2461,"BazaFilmow.sql")
dOcenaKobietyPonizej18(23456,2461,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

