% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dOcenaMezczyzniPonizej18.R
\name{dOcenaMezczyzniPonizej18}
\alias{dOcenaMezczyzniPonizej18}
\title{Liczenie odleglosci dla  ocen mezczyzn ponizej 18 lat}
\usage{
dOcenaMezczyzniPonizej18(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\value{
Funkcja zwraca wartosc liczbowa opisujaca odleglosc miedzy filmami.
}
\description{
Funkcja \code{dOcenaMezczyzniPonizej18} liczy odleglosc miedzy ocenami mezczyzn ponizej 18 lat dla
dwoch filmow z bazy
}
\details{
Funkcja wyciaga ID dwoch filmow z bazy. Jesli ktorys z nich nie ma ocen mezczyzn ponizej 18 lat
odleglosc miedzy nimi wynosi 0.5, jesli tak nie jest miara jest liczona zgodnie z funkcja PorownajOceny.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
dOcenaMezczyzniPonizej18(4000,2461,"BazaFilmow.sql")
dOcenaMezczyzniPonizej18(23456,2461,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

