% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dOcenaMezczyzniPowyzej45.R
\name{dOcenaMezczyzniPowyzej45}
\alias{dOcenaMezczyzniPowyzej45}
\title{Liczenie odleglosci dla  ocen mezczyzn od 45 lat wzwyz}
\usage{
dOcenaMezczyzniPowyzej45(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\value{
Funkcja zwraca wartosc liczbowa opisujaca odleglosc miedzy filmami.
}
\description{
Funkcja \code{dOcenaMezczyzniPowyzej45} liczy odleglosc miedzy ocenami mezczyzn od 45 lat wzwyz dla
dwoch filmow z bazy
}
\details{
Funkcja wyciaga ID dwoch filmow z bazy. Jesli ktorys z nich nie ma ocena mezczyzn od 45 lat wzwyz
odleglosc miedzy nimi wynosi 0.5, jesli tak nie jest miara jest liczona zgodnie z funkcja PorownajOceny.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
dOcenaMezczyzniPowyzej45(4000,2461,"BazaFilmow.sql")
dOcenaMezczyzniPowyzej45(23456,2461,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

