% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DodajTabeleOdleglosci.R
\name{DodajTabeleOdleglosci}
\alias{DodajTabeleOdleglosci}
\title{Dodaje pusta tabele odleglosci do bazy filmow}
\usage{
DodajTabeleOdleglosci(sciezkaDoBazy)
}
\arguments{
\item{sciezkaDoBazy}{sciezka do pliku z baza}

\item{nazwaTabeli}{nazwa dodawanej tabeli}
}
\value{
Funkcja zwraca wartosc TRUE jesli proces sie powiedzie.
}
\description{
Funkcja \code{DodajTabeleOdleglosci} do podanej bazy danych dodaje pusta
tabele przechowujaca odleglosci miedzy filmami
}
\details{
Funkcja dodaje pusta tabele z zainicjowanymi kolumnami:
IDFilm1, IDFilm2, dRokProdukcji, dOpisIMDB, dGatunki, dScenarzysci, dRezyserowie,
dKraje, dAktorzy, dAktorzyGlowni, dDAktorzyGwiazdy, dProducenci, dOcenaIMDB,
dOcenaMezczyzniPonizej18, dOcenaKobietyPonizej18, dOcenaMezczyzni18_29,
dOcenaKobiety18_29, dOcenaMezczyzni30_44, dOcenaKobiety30_44,
dOcenaMezczyzniPowyzej45, dOcenaKobietyPowyzej45.
W odpowiednich kolumnach mozna potem przetrzymywac odleglosci miedzy dwoma
filmami ze wzgledu na okreslona kategorie.
}
\examples{
#Nie wyoluj jesli nie masz takiej bazy
DodajTabeleOdleglosci("BazaFilmow.sql")
}
\author{
Karolina Wyszynska
}

