% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dScenarzysci.R
\name{dScenarzysci}
\alias{dScenarzysci}
\title{Liczenie odleglosci dla scenarzystow}
\usage{
dScenarzysci(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\value{
Funkcja zwraca wartosc liczbowa opisujaca odleglosc miedzy filmami.
}
\description{
Funkcja \code{dScenarzysci} liczy odleglosc miedzy scenarzystami dwoch filmow z bazy
}
\details{
Funkcja wyciaga ID dwoch filmow z bazy. Jesli ktorys z nich nie ma scenarzystow odleglosc
miedzy nimi wynosi 0.8, jesli tak nie jest miara jest liczona zgodnie z funkcja PorownajWektory.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
dScenarzysci(4000,2461,"BazaFilmow.sql")
dScenarzysci(23456,2461,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

