% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OdleglosciMiedzyFilmami.R
\name{OdleglosciMiedzyFilmami}
\alias{OdleglosciMiedzyFilmami}
\title{Wektor odleglosci miedzy filmami w bazie}
\usage{
OdleglosciMiedzyFilmami(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1,}{\code{IDFilm2} ID filmow dla ktorych chcemy liczyc odleglosci}

\item{sciezkaDoBazy}{sciezka do bazy, w ktorej przechowywane sa informacje o
filmach}
}
\value{
Wektor typu double o wartosciach odpowiadajacych wartosciom kolumn
z tabeli utworzonej przez \code{DodajTabeleOdleglosci}
}
\description{
Funkcja \code{OdleglosciMiedzyFilmami} oblicza wektor wszystkich odleglosci
miedzy dwoma filmami
}
\details{
Funkcja laczy sie z baza i korzystajac ze wszystkich funkcji typu d*
dolaczonych do pakietu oblicza wektor odleglosci miedzy dwoma filmami
}
\examples{
sciezkaDoBazy <- "PelnaBazaFilmow.sql" # Nie zadziala bez wczesniej wypelnionej bazy
OdleglosciMiedzyFilmami(23456, 4000, sciezkaDoBazy)
}
\author{
Karolina Wyszynska
}

