% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PobierzFilm.R
\name{PobierzFilm}
\alias{PobierzFilm}
\title{Wyszukiwanie informacji o filmach w bazie IMDb}
\usage{
\code{PobierzFilm(link)}
}
\arguments{
\item{link}{link do strony danego filmu w bazie IMDb
(struktura linku jak w przykladzie)}
}
\value{
Zwraca liste zlozona z ramek danych i wektorow przechowujacych
informacje o filmie
}
\description{
Funkcja \code{PobierzFilm} sciaga informacje o filmie z serisu IMDb
}
\details{
\code{PobierzFilm} przeszukuje strone danego filmu w bazie IMDb
szukajac przydatnych informacji o nim takich jak np. tytul, aktorzy, gatunek.
}
\examples{
PobierzFilm("http://www.imdb.com/title/tt0395699/")
}
\author{
Krzysztof Rudas
}

