% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PobierzLinki.R
\name{PobierzLinki}
\alias{PobierzLinki}
\title{Wyszukiwanie linkow do filmow z bazy IMDb}
\usage{
\code{PobierzLinki(gatunki = c("action", "adventure", "animation", "biography", "comedy", "crime",
                               "documentary", "drama", "family", "fantasy", "film_noir", "history",
                               "horror", "music", "musical", "mystery", "romance", "sci_fi",
                               "sport", "thriller", "war", "western"))}
}
\arguments{
\item{gatunki}{wektor typu character gatunkow zgodnych z dostepnymi w zaawansowanej wyszukiwarce serwisu IMDB}
}
\value{
Zwraca wartosc TRUE or FALSE w zaleznosci od tego czy wykonanie sie powiedzie czy nie
}
\description{
Funkcja \code{PobierzLinki} wysyla zapytanie do wyszukiwarki IMDB
w zaleznosci od wybranych gatunkow
}
\details{
\code{PobierzLinki} dla kazdego podanego gatunku wysyla zapytanie do wyszukiwarki
IMDB oraz zbiera stamtad liste linkow do wyszukanych filmow.
Wynik zapisuje w pliku oddzielnym dla kazdego gatunku.
}
\examples{
PobierzLinki("adventure")
}
\author{
Karolina Wyszynska
}

