% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PorownajOceny.R
\name{PorownajOceny}
\alias{PorownajOceny}
\title{Porownanie ocen dwoch filmow}
\usage{
PorownajOceny(o1, o2)
}
\arguments{
\item{o1,o2}{- oceny filmow jako wartosci typu numeric}
}
\value{
Zwraca liczbe z przedzialu [0, 1] wyznaczajaca odleglosc miedzy dwoma
filmami (liczona na podstawie oceny). Zero w przypadku, gdy filmy bardzo
podobne, jeden gdy tego podobienstwa w ogle nie ma.
}
\description{
Funkcja \code{PorownajOceny} porownuje oceny dwoch filmow
}
\details{
Oceny, ktore mozna przyznac poszczegolnym filmom zawieraja sie w
przedziale [1, 10], dlatego maksymalna odleglosc miedzy dwoma ocenami wynosi 9.
Stad funkcja \code{PorownajOceny} oblicza zwykla odleglosc miedzy miedzy dwiema
ocenami, a nastepnie normuje ja dzielac przez 9.
}
\examples{
ocena1 <- 8.6
ocena2 <- 7.5

PorownajOceny(ocena1, ocena2)
}
\author{
Karolina Wyszynska
}

