% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PorownajOpis.R
\name{PorownajOpis}
\alias{PorownajOpis}
\title{Porownanie opisow dwoch filmow}
\usage{
PorownajOpis(o1, o2)
}
\arguments{
\item{o1,o2}{- opisy filmow podane jako wektory typu character dlugosci 1}
}
\value{
Zwraca liczbe z przedzialu [0, 1] wyznaczajaca odleglosc miedzy dwoma
filmami (liczona na podstawie opisu). Zero w przypadku, gdy filmy bardzo
podobne, jeden gdy tego podobienstwa w ogle nie ma.
}
\description{
Funkcja \code{PorownajOpis} porownuje opisy dwoch filmow (w jezyku angielskim)
}
\details{
Funkcja korzystajac z pakietow \code{tm} i \code{SnowballC}
oczyszcza tekst (musi byc w jezyku angielskim). Tak przetworzone opisy sa
rozbijane na dwa wektory slow. Wektory te sa nastepnie porownywane przy uzyciu
funkcji \code{PorownajWektory}
}
\examples{
opis1 <- "This time Ellena falls in love with Richard and they live happily
togehter ever after."
opis2 <- "Cathrene get engaged with Tom and they live happily together."

PorownajOpis(opis1, opis2)
}
\author{
Karolina Wyszynska
}

