% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PorownajRok.R
\name{PorownajRok}
\alias{PorownajRok}
\title{Porownanie lat produkcji dwoch filmow}
\usage{
PorownajRok(r1, r2)
}
\arguments{
\item{r1,r2}{- rok produkcji poszczegolnych filmow jako wartosci typu numeric}
}
\value{
Zwraca liczbe z przedzialu [0, 1] wyznaczajaca odleglosc miedzy dwoma
filmami (liczona na podstawie roku produkcji). Zero w przypadku, gdy filmy bardzo
podobne, jeden gdy tego podobienstwa w ogle nie ma.
}
\description{
Funkcja \code{PorownajRok} porownuje lata produkcji dwoch filmow
}
\details{
Funkcja oblicza zwykla odleglosc miedzy dwoma latami, a nastepnie
dzieli ja przez 30. Uznalismy, ze 30 lat to najdluzej kiedy filmy moga byc
jeszcze uznane za podobne, dlatego dla wiekszych odleglosci funkcja zwraca 1.
}
\examples{
rok1 <- 1992
rok2 <- 2000

PorownajRok(rok1, rok2)

rok1 <- 1960
rok2 <- 2034

PorownajRok(rok1, rok2)
}
\author{
Karolina Wyszynska
}

