% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PorownajWektory.R
\name{PorownajWektory}
\alias{PorownajWektory}
\title{Porownanie dwa wektory}
\usage{
PorownajWektory(w1, w2)
}
\arguments{
\item{w1,}{w2 - dwa wektory typu character}
}
\value{
Zwraca liczbe z przedzialu [0, 1] wyznaczajaca odleglosc miedzy dwoma
wektorami. Zero w przypadku, gdy wektory bardzo
podobne, jeden gdy tego podobienstwa w ogle nie ma.
}
\description{
Funkcja \code{PorownajWektory} bada podobienstwo dwoch wektorow
}
\details{
Funkcja oblicza dlugosci obu wektorow, a nastepnie wyciaga minimum z
tych wielkosci i liczby 6 (uznalismy, ze 6 wspolnych aktorow to wystarczajace podobienstwo). Nastepnie
okresla dlugosc przeciecia obu wektorow. Normuje ja przez wczesniej wyliczona
wielkosc i oblicza odleglosc od 1.
}
\examples{
wektor1 <- c("Adam Sandler","John Turturro","Emmanuelle Chriqui")
wektor2 <- c("Adam Sandler", "Drew Barrymore", "Rob Schneider")

PorownajWektory(wektor1, wektor2)
}
\author{
Karolina Wyszynska
}

