% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PorownajWektory1.R
\name{PorownajWektory1}
\alias{PorownajWektory1}
\title{Porownanie dwa wektory}
\usage{
PorownajWektory1(w1, w2)
}
\arguments{
\item{w1,}{w2 - dwa wektory typu character}
}
\value{
Zwraca liczbe z przedzialu [0, 1] wyznaczajaca odleglosc miedzy dwoma
wektorami. Zero w przypadku, gdy wektory bardzo
podobne, jeden gdy tego podobienstwa w ogole nie ma.
}
\description{
Funkcja \code{PorownajWektory1} bada podobienstwo dwoch wektorow
}
\details{
Funkcja oblicza dlugosci obu wektorow, najpierw liczymy sume (w sensie mnogosciowym)
dlugosci wektorów, a takze dlugosc przeciecia wektorow. nastepnie dzielimy dlugosc przeciecia przez
sume dlugosci wektorow i to wszystko odejmujemy od 1. Ta miare zastosujemy dla gatunkow i krajow.
Stosujemy ja aby zabezpieczyc sie przed przypadkiem gdy jeden film ma np. tylko gatunek Drama, a drugi
duzo gatunkow (w tym Drama). Niezaleznie od tego ile gatunkow bedziemy mieli w drugim filmie
wczesniejsza funkcja (PorownajWektory) zwracalaby zawsze 0.
}
\examples{
wektor1 <- c("Drama","Crime")
wektor2 <- c("Drama", "Adventure")

PorownajWektory1(wektor1, wektor2)
}
\author{
Karolina Wyszynska
}

