% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RysujMape.R
\name{RysujMape}
\alias{RysujMape}
\title{Rysowanie mapy odleglosci miedzy filmami}
\usage{
RysujMape(macierzOdleglosci)
}
\arguments{
\item{macierzOdleglosci}{- macierz odleglosci}
}
\description{
Funkcja \code{RysujMape} rysuje mape odleglosci pomiedzy filmami
}
\details{
Funkcja przyjmuje jak argument macierz odleglosci generowana przez
funkcje \code{StworzMacierzOdleglosci}. Na jej podstawie rysuje mape odleglosci
}
\examples{
IDFilm <- 67484
wektorWspolczynnikow <- WektorWspolczynnikow(c("Oceny", "Producenci"), "K", 22)
sciezkaDoBazy<-paste0(getwd(),Baza_Filmow.sql)
RysujMape(StworzMacierzOdleglosci(IDFilm, wektorWspolczynnikow, sciezkaDoBazy))
}
\author{
Krzysztof Rudas
}

