% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SprawdzCzywBazie.R
\name{SprawdzCzywBazie}
\alias{SprawdzCzywBazie}
\title{Sprawdzenie, czy dane ID jest w bazie.}
\usage{
SprawdzCzywBazie(IDFilm1, IDFilm2, baza_filmow)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{baza_filmow}{- polaczenie z bazą filmow}
}
\value{
Funkcja zwraca wartosc TRUE jesli obydwa ID znajduja sie w bazie, jesli nie to mamy wartosc FALSE.
}
\description{
Funkcja \code{SprawdzCzywBazie} sprawdza, czy para filmow o podanych przez użytkownika ID
jest w bazie filmow.
}
\details{
Funkcja korzysta juz z polaczenia z baza filmow. W tabeli Filmy, sprawdza czy istnieja zadane
w parametrach ID filmow.
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
SprawdzCzywBazie(4000,2461,dbConnect(SQLite(),dbname="BazaFilmow.sql"))
SprawdzCzywBazie(900,2461,dbConnect(SQLite(),dbname="BazaFilmow.sql"))
}
\author{
Krzysztof Rudas
}

