% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StworzBaze.R
\name{StworzBaze}
\alias{StworzBaze}
\title{Tworzenie bazy przechowujacej informacje o filmach}
\usage{
StworzBaze(sciezka_do_bazy)
}
\arguments{
\item{sciezka_do_bazy}{sciezka do nowotworzonej bazy}
}
\value{
Zwraca wartosc TRUE swiadczaca o utworzeniu relacyjnej bazy danych o
zadanych tabelach.
}
\description{
Funkcja \code{StworzBaze} tworzy relacyjna baze danych przechowujaca
informacje o filamch
}
\details{
\code{StworzBaze(sciezka_do_bazy)}

\code{StworzBaze} generuje tabele przechowujace informacje min.
o tytulach, gatunkach, aktorach. Oprocz tabeli przechwujacych dane,
konstruowane sa rowniez tabele wiele do wielu, ktore umozliwiaja powiazanie
miedzy tabelami przechowujacymi rozne informacje.
}
\examples{
sciezka <- getwd()
StworzBaze(paste0(sciezka, "bazaFilmow.sql"))
}
\author{
Krzysztof Rudas

Karolina Wyszynska
}

