% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StworzMacierzOdleglosci.R
\name{StworzMacierzOdleglosci}
\alias{StworzMacierzOdleglosci}
\title{Liczenie macierzy  odleglosci dla zadanego filmu}
\usage{
StworzMacierzOdleglosci(IDFilm, wektorWspolczynnikow, sciezkaDoBazy)
}
\arguments{
\item{IDFilm}{ID filmu na podstawie, ktorego chcemy wybrac 10 najblizszych
mu filmow}

\item{wektorWspolczynnikow}{wektor zwracany przez funkcje \code{WektorWspolczynnikow}
okreslajacy wage poszczegolnych rzeczy dla uzytkownika}

\item{sciezkaDoBazy}{sciezka do bazy, w ktorej przechowujemy informacje o filmach}
}
\value{
Macierz odleglosci z tytulami filmow w wierszach i kolumnach
}
\description{
Funkcja \code{StworzMacierzOdleglosci} zwraca macierz odleglosci 10 filmow
najblizszych wybranemu filmowi.
}
\details{
Funkcja operujac na bazie danych znajduje 10 najblizszych filmow dla
IDFilm oraz oblicza macierz odleglosci pomiedzy tymi 11 (lacznie z IDFilm) filmami
}
\examples{
sciezkaDoBazy <- "BazaFilmow.sql"
IDFilm <- 126029
wektorWspolczynnikow <- WektorWspolczynnikow(c("Oceny", "Producenci"), "K", 22)
StworzMacierzOdleglosci(IDFilm, wektorWspolczynnikow, sciezkaDoBazy)
}
\author{
Karolina Wyszynska
}

