% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UzupelnijBaze.R
\name{UzupelnijBaze}
\alias{UzupelnijBaze}
\title{Uzupelnia baze danych}
\usage{
\code{UzupelnijBaze(listaLinkow, sciezkaDoBazy)}
}
\arguments{
\item{listaLinkow}{wektor typu character zawierajacy linki do filmow,
ktore chcemy wstawic do bazy}

\item{sciezkaDoBazy}{sciezka do pliku z baza danych}
}
\description{
Funkcja \code{UzupelnijBaze} jest 'ladnym opakowaniem' na funkcje
\code{WstawFilmDoBazy}. Przyjmuje wektor linkow i po koleji
umieszcza je w bazie za pomoca wczesniej wspomnianiej funkcji
}
\examples{
UzupelnijBaze(c("http://www.imdb.com/title/tt0395699/",
               "http://www.imdb.com/title/tt0006033/"),
               "nazwaBazy.sql")
}
\author{
Karolina Wyszynska
}

