% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WektorWspolczynnikow.R
\name{WektorWspolczynnikow}
\alias{WektorWspolczynnikow}
\title{Wagi odpowiadajace preferencjom uzytkownika}
\usage{
WektorWspolczynnikow(coWazne, plec, wiek)
}
\arguments{
\item{coWazne}{dwuelementowy wektor typu character podajacy dwie najwazniejsze
dla uzytkownika cechy filmu. Dostepne wartosci to: Rok, Fabula, Gatunki,
Scenarzysci, Rezyserowie, Kraje produkcji, Aktorzy, Gwiazdy i Producenci.}

\item{plec}{plec uzytkownika jako wartosc "K" jesli kobieta lub "M" jesli mezczyzna}

\item{wiek}{wiek uzytkownika podany w latach}
}
\value{
Wektor typu double wag dla poszczegolnych kolumn tabeli Odleglosci
}
\description{
Funkcja \code{WektorWspolczynnikow} oblicza wektor wag z uwzglednieniem
preferencji uzytkownika
}
\details{
Funkcja oblicza wagi dla poszczegolnych cech filmow. Dwum wyroznionym
cechom przyporzadkowywuje wartosc 0.23, reszcie wartosc 0.06 . Dodatkowo
ceche ocena oblicza jako srednia artmetyczna oceny glownej oraz tej
odpowiadajacej wiekowi i plci uzytkowika
}
\examples{
wektorWspolczynnikow <- WektorWspolczynnikow(c("Oceny", "Producenci"), "K", 22)
}
\author{
Karolina Wyszynska
}

