% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WstawFilmDoBazy.R
\name{WstawFilmDoBazy}
\alias{WstawFilmDoBazy}
\title{Wstawianie informacji o filmie do bazy}
\usage{
\code{WstawFilmDoBazy(linkDoFilmu, sciezkaDoBazy)}
}
\arguments{
\item{linkDoFilmu}{link do interesujacego nas filmu w serwisie IMBD}

\item{sciezkaDoBazy}{sciezka do bazy znajdujacej sie w pliku}
}
\value{
Zwraca wartosc TRUE jesli zapis sie powiodl, FALSE
wraz z komentarzem w przeciwnym przypadku.
}
\description{
Funkcja \code{WstawFilmDoBazy} sciaga informacje o filmie
i wstawia je do odpowiedniej bazy danych
}
\details{
\code{WstawFilmDoBazy} sciagajac informacje o podanym filmie
korzysta przy tym z funkcji \code{PobierzFilm}. Wartosci pobrane z wyniku
lokuje w bazie analogicznej do \code{StworzBaze}. Funkcja sama nie tworzy bazy.
Funkcja nie przyjmuje filmow krotkometrazowych.
}
\examples{
sciezka <- getwd()
StworzBaze(paste0(sciezka, "bazaFilmow.sql"))
WstawFilmDoBazy("http://www.imdb.com/title/tt0395699/", paste0(sciezka, "bazaFilmow.sql"))
}

