% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WstawMiaryDoCSV.R
\name{WstawMiaryDoCSV}
\alias{WstawMiaryDoCSV}
\title{Wrzucanie wyliczonych miar do pliku CSV}
\usage{
WstawMiaryDoCSV(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1}{- id pierwszego filmu}

\item{IDFilm2}{- id drugiego filmu}

\item{sciezkaDoBazy}{- sciezka do bazy filmow}
}
\description{
Funkcja \code{WstawMiaryDoCSV} liczy wektory miar dla filmow i wrzuca je do pliku CSV
}
\details{
Funkcja jest rozszerzeniem funkcji \code{WstawOdlegloscDoCSV}. Dla zadanego ID filmu
pierwszego i filmu drugiego zlicza wektory odleglosci i zapisuje je do csv w nastepujacy sposob:
Najpierw dla wszystkich par filmow, gdzie pierwszym filmem jest film o ID rownym IDFilm1, a drugim
filmy o ID wiekszym od IDFilm2. Pozniej IDFilm1 zmienia sie na nastepne i liczone sa miary dla nowego
IDFilm1 i IDFilm2 wiekszych od IDFilm1, nastepnie znow IDFilm1 staje sie nastepnym w kolejnosci i proces
sie powtarza. (Nalezy podawac IDFilm2 wieksze lub rowne od IDFilm1).
}
\examples{
#Nie wywoluj jesli nie masz takiej bazy
WstawMiaryDoCSV(2461,4000,"BazaFilmow.sql")
WstawMiaryDoCSV(2461,23456,"BazaFilmow.sql")
}
\author{
Krzysztof Rudas
}

