% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WstawOdlegloscDoBazy.R
\name{WstawOdlegloscDoBazy}
\alias{WstawOdlegloscDoBazy}
\title{Zamieszcza wektor odleglosci miedzy filmami w bazie}
\usage{
WstawOdlegloscDoBazy(IDFilm1, IDFilm2, sciezkaDoBazy)
}
\arguments{
\item{IDFilm1,}{\code{IDFilm2} ID filmow dla ktorych chcemy liczyc odleglosci}

\item{sciezkaDoBazy}{sciezka do bazy, w ktorej przechowywane sa informacje o
filmach}
}
\value{
TRUE jesli proces sie powiedzie, FALSE w przeciwnym przypadku
}
\description{
Funkcja \code{WstawOdlegloscDoBazy} umieszcza wektor wszystkich odleglosci
miedzy dwoma filmami w bazie.
}
\details{
Funkcja laczy sie z baza i korzystajac z funkcji
\code{OdleglosciMiedzyFilmami} oblicza wektor odleglosci oraz umieszcza go
w bazie
}
\examples{
sciezkaDoBazy <- "PelnaBazaFilmow.sql" # Nie zadziala bez wczesniej wypelnionej bazy
WstawOdlegloscDoBazy(23456, 4000, sciezkaDoBazy)
}
\author{
Karolina Wyszynska
}

