% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WstawOdlegloscDoCSV.R
\name{WstawOdlegloscDoCSV}
\alias{WstawOdlegloscDoCSV}
\title{Zamieszcza wektor odleglosci miedzy filmami w pliku CSV}
\usage{
WstawOdlegloscDoCSV(IDFilm1, IDFilm2, sciezkaDoBazy, sciezkaDoCSV)
}
\arguments{
\item{IDFilm1,}{\code{IDFilm2} ID filmow dla ktorych chcemy liczyc odleglosci}

\item{sciezkaDoBazy}{sciezka do bazy, w ktorej przechowywane sa informacje o
filmach}

\item{sciezkaDoCSV}{sciezka do pliku typu '.csv', w ktorym chcemy trzymac dane}
}
\value{
TRUE jesli proces sie powiedzie, FALSE w przeciwnym przypadku
}
\description{
Funkcja \code{WstawOdlegloscDoCSV} umieszcza wektor wszystkich odleglosci
miedzy dwoma filmami w wybranym pliku typu '.csv'.
}
\details{
Funkcja laczy sie z baza i korzystajac z funkcji
\code{OdleglosciMiedzyFilmami} oblicza wektor odleglosci oraz umieszcza go
w pliku CSV
}
\examples{
sciezkaDoCSV <- "Odleglosci.csv"
if(!file.exists(sciezkaDoCSV)){file.create(sciezkaDoCSV)}
sciezkaDoBazy <- "PelnaBazaFilmow.sql" # Nie zadziala bez wczesniej wypelnionej bazy
WstawOdlegloscDoCSV(23456, 4000, sciezkaDoBazy)
}
\author{
Karolina Wyszynska
}

