% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ZCSVDoBazy.R
\name{ZCSVDoBazy}
\alias{ZCSVDoBazy}
\title{Wczytywanie odleglosci z CSV do bazy}
\usage{
ZCSVDoBazy(sciezkaDoBazy, sciezkaDoCSV, wczytujCo, nazwaTabeli)
}
\arguments{
\item{sciezkaDoBazy}{sciezka do bazy, w ktorej trzymamy informacje o filmach}

\item{sciezkaDoCSV}{sciezka do pliku '.csv', w ktyrm tymczasowo przechowujemy dane}

\item{wczytujCo}{ile wierszy z pliku chcemy wczytywac na raz}

\item{nazwaTabeli}{do jakiej tabeli w bazie chcemy to wczytac (musi miec kolumny)
analogiczne jak tabela ze \code{DodajTabeleOdleglosci}}
}
\value{
Zwraca wartosc TRUE
}
\description{
Funkcja \code{ZCSVDoBazy} zczytuje wektory odleglosci pomiedzy filmami z pliku
CSV do bazy danych
}
\examples{
sciezkaDoBazy <- "BazaFilmow.sql"
sciezkaDoCSV <- "PrzechowajOdleglosci.csv"
ZCSVDoBazy(sciezkaDoBazy, sciezkaDoCSV, wczytujCo=100, "Odleglosci")
}
\author{
Karolina Wyszynska
}

